Move logging to lib
