+++ scripts/lights.js
@@ -11,7 +11,7 @@
 const _ = require('lodash');
 const fetch = require('node-fetch');
 const openFaas = require('../lib/openfaas');
-const logger = require('./log');
+const logger = require('../lib/log');
 
 function sendCommand(command) {
   payload = {
