+++ scripts/nummer.js
@@ -6,7 +6,7 @@
 
 const _ = require('lodash');
 const members = require('../lib/members');
-const logger = require('./log');
+const logger = require('../lib/log');
 
 module.exports = robot => {
   robot.respond(/nummer/i, msg => {
