+++ scripts/tail.js
@@ -4,7 +4,7 @@
 // Commands
 //   hubot tail [options] <command>
 
-const logger = require('./log');
+const logger = require('../lib/log');
 
 module.exports = robot => {
   robot.respond(/tail (.*)?/i, msg => {
